Create fibonacci series 87(java)
An e-commerce website wishes to find the lucky customer who will be eligible for full value cash back. For this purpose, a number N is fed to the system. It will return another number that is calculated by an algorithm. In the algorithm, a sequence is generated, in which each number is the sum of the two proceeding numbers. Initially the sequence will have two 1’s in it. The system will return the Nth number from the generated sequence which is treated as the order ID. The lucky customer will be the one who has placed that order. Write an algorithm to help the website find the lucky customer.

Input Format

The input consists of an integer token, representing the number fed to the system (N).

Constraints

1<=n<=10^7

Output Format

Print an integer representing the order ID of the lucky customer.

Sample Input 0

8
Sample Output 0

21
Explanation 0

The sequence generated by the system will be 1,1,2,3,5,8,13,21. The 8th number in the sequence is 21. The lucky customer is the one who has placed the order with order ID 21.

Sample Input 1

2
Sample Output 1

1
Sample Input 2

10
Sample Output 2

55